Add button to hide/show all snippets
One complaint was that it takes too many clicks to judge a set of
results. This should reduce the number of clicks it takes, at the
expense of screen real estate.  Might be worthwhile to have different
default states depending on if the user is on mobile or desktop, but
havn't decided.

Change-Id: Ibe8541b35f9f0968c927ddf0bee73dad1814a504