Keep track of snippet highlighting
It might be easier for judges if we make the grading interface look more
like a search result page they are used to. Part of that is highlighting
relevant pieces of the snippet.
Rather than try and integrate our own highlighter, which would need to
know about spelling corrections and such, keep track of what was
highlighted by the search engine that provided the snippet and
re-highlight when sending out the scoring form.

Change-Id: I52bf2129beba07c76557604fcd7c2daa7ad84f60