Add a bare bones form rendering template for cards
Jan will be working on a card based interface for discernatron. Setup
a simple template that outputs the various data used in scoring as
a json structure, along with a form full of hidden inputs that can
be updated with relevant scores.

This isn't wired in very well yet, it has to be explicitly enabled
by attaching ?cards=1 to a scoring page, such as

  http://192.168.33.10/query/id/2?cards=1

Can revisit later, perhaps with some sort of per-user option or just
the new default, when we make that decision.

Change-Id: I4d7b8e0e386fbf3f0fadd65fefbb94cadf30eaf2