Change scoring queue priority to 1,1,3,4,5
The scoring queue was, by default, always assigning scoring priority
as 1,2,3,4,5. This has the side effect of requiring every query to
receive one set of scores before looping back arround and getting a
second score. After some analysis of the results we are now fairly
certain we want to get at least 2 sets of scores for each query.

This removes the scoring-queue:update console method, it was never
used and would take some bit of refactoring to make work with the
new method. Rather than figure out how, just remove it.

Change-Id: Ic440b62ec1bb102e7b526e58a8f477b6e1212f0f