Fix issue with dropping card back on top deck
When dragging a card from the top deck and dropping it back there is an
issue that temporarily breaks the deck until some odd user interaction
happen. This looks to be fallout of the change that fixed handling of
moving cards back into the top deck. Checking if the card is already
the currentCard and handling it slightly differently seems to fix the
issue.

Change-Id: I7a4a2d46676bfc6fd80e09c8df2fa48805729e10