Show an error page when visiting un-imported query
If a user adjusts the url, or somehow we give a bad url, to a query that
is not yet imported the scoring page is rendered but no results are
provided. Instead of rendering the results page give a more specific
error page that links them to a new query to grade.

Change-Id: If7add588dd2015a7b33f2c8822214876aa30b624