Don't re-request user aproval via oauth
Using the authenticate, instead of the authorize, endpoint allows
logins to pass through mediawiki without a confirmation step as long as
the user has approved the application at some time in the past.

The user will still have to re-login whenever their session times out,
which on the discernatron servers is 6 hours.

Change-Id: I5d836de4238fab549cfc52ee1c16bf5d64cf6cf1