Change items inserted in scoring queue to match reality
We have been inserting a bunch of items into the scoring queue that we
don't actually expect to ever be done. Rather than injecting the 5 times
we 'hope' the query will be scored, insert 2 to match what we actually
expect to happen. This should make it easier to write the code for
injecting new items to the queue for queries without enough agreement.

Change-Id: I2fa781f1054f55b618698352d998e95ed59a0e15