Track reliability of query scores
There can be problems where we have a large disagreement between users
about what should or should not be relevant for a query. Try to resolve
that by adding the query back into the scoring queue to get a couple
more users to look at it.

Also adds a 'reliable' column to the scores, so that relevance forge can
filter on it and only pull in scores that have been deemed reliable.
This utilizes a standard measure of agreement between multiple-coders
to evaluate if the different grades are in agreement, although I'm
uncertain of what thresholds to use. The "standard" thresholds are 0.8
or 0.667, but these end up flagging quite a few of the current scores.

Bug: T146189
Change-Id: I36091436acb81b343a14ca0801ce77924055a654